requirements.txt
appium-python-client>=2.0.0
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
PyYAML>=5.4.0
pytest>=7.0.0
allure-pytest>=2.12.0